added settings interactor
